Add Writer class
The idea behind this class is to allow the user to write data sequentially. The problem with the existing code is that all the data must reside in memory before it can be written with write(). For large datasets or those that are being acquired in real time (e.g. from a sensor), this approach doesn't work.

The Writer class provides much the same interface as the writer objects returned from csv.writer. There is also a context manager for automatically closing the file handle.

Closes #13.